% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ensure_trailing_slash.R
\name{ensure_trailing_slash}
\alias{ensure_trailing_slash}
\title{ensure trailing slash function}
\usage{
ensure_trailing_slash(path)
}
\arguments{
\item{path}{Input path (or filestring)}
}
\description{
takes an input (path) and checks for a final slash
}

