% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fml.R
\name{fml}
\alias{fml}
\title{fml function
based on
http://stackoverflow.com/questions/21718711/writing-a-shortcut-function-in-r-for-rmlist-ls
http://stackoverflow.com/questions/14260340/function-to-clear-the-console-in-r}
\usage{
fml()
}
\description{
fml function
based on
http://stackoverflow.com/questions/21718711/writing-a-shortcut-function-in-r-for-rmlist-ls
http://stackoverflow.com/questions/14260340/function-to-clear-the-console-in-r
}

