% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graph_stacked_by_output_position.R
\name{graph_stacked_by_output_position}
\alias{graph_stacked_by_output_position}
\title{Create stacked bar graphs}
\usage{
graph_stacked_by_output_position(data)
}
\arguments{
\item{data:}{dataframe of summarized data}
}
\description{
This function takes a dataframe (data),to create a graph of responses stacked by type and split into bars by sublist across output positions.
To get graphs to print, must wrap function call in print().
}

